Moved and refactored UriUtility tests
